requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-autoslug-iplweb==1.9.4.dev0  # via ndh
django-bootstrap4==0.0.6  # via ndh
django-filter==1.1.0
django-tables2==1.21.1
django==2.0.3             # via django-tables2, ndh
ndh==3.3.0
pytz==2018.3              # via django
